+++ pkg/distributor/distributor_test.go
@@ -2472,7 +2472,7 @@ func TestDistributor_IngestionIsControlledByForwarder(t *testing.T) {
 			var forwardReqCnt atomic.Uint32
 			forwardReqCallback := func() { forwardReqCnt.Inc() }
 			getForwarder := func() forwarding.Forwarder {
-				return forwarding.NewMockForwarder(tc.ingestSample, forwardReqCallback)
+				return newMockForwarder(tc.ingestSample, forwardReqCallback)
 			}
 
 			distributors, ingesters, regs := prepare(t, prepConfig{
@@ -3547,6 +3547,47 @@ func (i *mockIngester) UserStats(ctx context.Context, in *client.UserStatsReques
 		RuleIngestionRate: 0,
 	}, nil
 }
+
+type mockForwarder struct {
+	ingest bool
+
+	// Optional callback to run in place of the actual forwarding request.
+	forwardReqCallback func()
+}
+
+func newMockForwarder(ingest bool, forwardReqCallback func()) forwarding.Forwarder {
+	return &mockForwarder{
+		ingest:             ingest,
+		forwardReqCallback: forwardReqCallback,
+	}
+}
+
+func (m *mockForwarder) Forward(ctx context.Context, forwardingRules validation.ForwardingRules, ts []mimirpb.PreallocTimeseries) (forwarding.TimeseriesCounts, []mimirpb.PreallocTimeseries, chan error) {
+	errCh := make(chan error)
+
+	go func() {
+		defer close(errCh)
+
+		if m.forwardReqCallback != nil {
+			m.forwardReqCallback()
+		}
+	}()
+
+	var notIngestedCounts forwarding.TimeseriesCounts
+
+	if m.ingest {
+		return notIngestedCounts, ts, errCh
+	}
+
+	for _, t := range ts {
+		notIngestedCounts.SampleCount += len(t.Samples)
+		notIngestedCounts.ExemplarCount += len(t.Exemplars)
+	}
+
+	return notIngestedCounts, nil, errCh
+}
+
+func (m *mockForwarder) Stop() {}
 
 func match(labels []mimirpb.LabelAdapter, matchers []*labels.Matcher) bool {
 outer:
